Change background colour to red when the time gets low
