+++ shared/chat/conversation/messages/message-popup/exploding/index.js
@@ -2,7 +2,7 @@
 import * as React from 'react'
 import {Avatar, Box2, FloatingMenu, Icon, Text} from '../../../../../common-adapters/'
 import {collapseStyles, globalColors, globalMargins, isMobile, platformStyles} from '../../../../../styles'
-import {formatTimeForPopup, formatTimeForRevoked} from '../../../../../util/timestamp'
+import {formatTimeForPopup, formatTimeForRevoked, secondsToDHMS} from '../../../../../util/timestamp'
 import {PopupHeaderText} from '../../../../../common-adapters/popup-menu'
 import type {DeviceType} from '../../../../../constants/types/devices'
 import type {Position} from '../../../../../common-adapters/relative-popup-hoc'
@@ -27,16 +27,6 @@ type Props = {
 type State = {
   secondsLeft: number,
 }
-const secondsToFormat = (seconds: number) => {
-  let mins = Math.floor(seconds / 60)
-  let hours = Math.floor(mins / 60)
-  let days = Math.floor(hours / 24)
-  let secs = seconds % 60
-  hours = hours % 24
-  mins = mins % 60
-
-  return `${days}d ${hours}h ${mins}m ${secs}s`
-}
 
 class ExplodingPopupHeader extends React.Component<Props> {
   state: State = {
@@ -49,8 +39,8 @@ class ExplodingPopupHeader extends React.Component<Props> {
   }
 
   componentWillMount() {
-    console.warn('in cwm')
     this.timer = setInterval(() => this.tick(), 1000)
+    this.tick()
   }
 
   componentWillUnmount() {
@@ -59,12 +49,12 @@ class ExplodingPopupHeader extends React.Component<Props> {
 
   tick() {
     this.setState({
-      secondsLeft: this.props.explodesAt - (Math.floor(Date.now() / 1000)),
+      secondsLeft: this.props.explodesAt - Math.floor(Date.now() / 1000),
     })
   }
 
   render() {
-    const {author, deviceName, deviceRevokedAt, explodesAt, timestamp, yourMessage} = this.props
+    const {author, deviceName, deviceRevokedAt, timestamp, yourMessage} = this.props
     const whoRevoked = yourMessage ? 'You' : author
     return (
       <Box2 direction="vertical" fullWidth={true} style={{alignItems: 'center'}}>
@@ -81,7 +71,7 @@ class ExplodingPopupHeader extends React.Component<Props> {
           <Text type="BodySmall" style={{color: globalColors.black_40}}>
             by
           </Text>
-          <Avatar style={{marginLeft: globalMargins.xtiny, marginRight: globalMargins.xtiny}} username={author} size={12} />
+          <Avatar style={styleAvatar} username={author} size={12} />
           <Text type="BodySmallItalic" style={{color: globalColors.black_60}}>
             {author}
           </Text>
@@ -105,8 +95,20 @@ class ExplodingPopupHeader extends React.Component<Props> {
             {whoRevoked} revoked this device on {formatTimeForRevoked(deviceRevokedAt)}.
           </PopupHeaderText>
         )}
-        <Box2 direction="vertical" gap="xsmall" fullWidth={true} gapEnd={true} gapStart={true} style={{backgroundColor: globalColors.black, marginTop: globalMargins.tiny}}>
-          <Text style={{color: globalColors.white, textAlign: 'center'}} type="BodySemibold">{secondsToFormat(this.state.secondsLeft)}</Text>
+        <Box2
+          direction="vertical"
+          gap="xsmall"
+          fullWidth={true}
+          gapEnd={true}
+          gapStart={true}
+          style={{
+            backgroundColor: this.state.secondsLeft < oneHourInSecs ? globalColors.red : globalColors.black,
+            marginTop: globalMargins.tiny,
+          }}
+        >
+          <Text style={{color: globalColors.white, textAlign: 'center'}} type="BodySemibold">
+            {secondsToDHMS(this.state.secondsLeft)}
+          </Text>
         </Box2>
       </Box2>
     )
@@ -156,6 +158,8 @@ const ExplodingPopupMenu = (props: Props) => {
   )
 }
 
+const oneHourInSecs = 60 * 60
+
 const stylePopup = platformStyles({
   common: {
     overflow: 'visible',
@@ -167,6 +171,11 @@ const stylePopup = platformStyles({
     width: '100%',
   },
 })
+
+const styleAvatar = {
+  marginLeft: globalMargins.tiny,
+  marginRight: globalMargins.xtiny,
+}
 
 const styleRevokedAt = {
   borderBottomLeftRadius: 3,
